Add installer services and sheet loader for game database management
- Implemented `installer_services.py` to handle installation from source folders, archive extraction, and downloading files.
- Added functions for installing OptiPatcher and Unreal5 from URLs, including handling of dinput8.dll for REFramework.
- Created `sheet_loader.py` to load game database from public Google Sheets, including support for various column mappings and settings.
- Introduced error handling and logging for download and extraction processes.
